dashboard/requirements.txt
dash==2.17.0
dash-bootstrap-components==1.6.0
plotly==5.22.0
pandas==2.2.2
psycopg2-binary==2.9.9
geopandas==0.14.4
Shapely==2.0.4
SQLAlchemy==2.0.30
gunicorn==22.0.0
